requirements.txt
ultralytics>=8.0.0
streamlit>=1.20.0
opencv-python
Pillow
